Extract `getResolvedSize` into a helper
