+++ packages/umi-serializer-beet/src/createBeetSerializer.ts
@@ -1,5 +1,4 @@
 import {
-  ArrayLikeSerializerSize,
   ArraySerializerOptions,
   BoolSerializerOptions,
   DataEnum,
@@ -32,6 +31,7 @@ import {
   DeserializingEmptyBufferError,
   NotEnoughBytesError,
 } from './errors';
+import { getResolvedSize } from './getResolvedSize';
 import { getSizeDescription } from './getSizeDescription';
 import { getSizeFromChildren } from './getSizeFromChildren';
 import { getSizePrefix } from './getSizePrefix';
@@ -653,38 +653,4 @@ function maxSerializerSizes(sizes: (number | null)[]): number | null {
     (all, size) => (all === null || size === null ? null : Math.max(all, size)),
     0 as number | null
   );
-}
-
-function getResolvedSize(
-  size: ArrayLikeSerializerSize,
-  childrenSizes: (number | null)[],
-  bytes: Uint8Array,
-  offset: number
-): [number | bigint, number] {
-  if (typeof size === 'number') {
-    return [size, offset];
-  }
-
-  if (typeof size === 'object') {
-    return size.deserialize(bytes, offset);
-  }
-
-  if (size === 'remainder') {
-    const childrenSize = sumSerializerSizes(childrenSizes);
-    if (childrenSize === null) {
-      throw new BeetSerializerError(
-        'Serializers of "remainder" size must have fixed-size items.'
-      );
-    }
-    const remainder = bytes.slice(offset).length;
-    if (remainder % childrenSize !== 0) {
-      throw new BeetSerializerError(
-        `Serializers of "remainder" size must have a remainder that is a multiple of its item size. ` +
-          `Got ${remainder} bytes remaining and ${childrenSize} bytes per item.`
-      );
-    }
-    return [remainder / childrenSize, offset];
-  }
-
-  throw new BeetSerializerError(`Unknown size type: ${JSON.stringify(size)}.`);
 }
